requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
attrs==24.2.0
babel==2.16.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
docutils==0.21.2
email-validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imagesize==1.4.1
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
libsql-client==0.3.1
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==2.0.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-core==2.20.1
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1
requests==2.32.3
rich==13.7.1
scikit-learn==1.4.2
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sphinx==8.0.2
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlalchemy==2.0.31
sqlalchemy-libsql==0.1.0
starlette==0.37.2
threadpoolctl==3.5.0
typer==0.12.3
typing-extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
